Initial commit
All what it needs to be installed and run is present, code is commented and can be easily changed. To have it autorun go to run (Win-R) and type "shell:sendto" without the quotation marks and drop a shortcut of the script to that folder, it will auto start with windows.